Use recommended module layout
