+++ domain/lib/domain.dart
@@ -1,20 +1,1 @@
-class Counter {
-  int _value;
-
-  Counter([int? value]) : _value = value ?? 0;
-
-  int get value => _value;
-
-  void increment() {
-    _value++;
-  }
-
-  bool tryDecrement() {
-    if (value == 0) {
-      return false;
-    }
-
-    _value--;
-    return true;
-  }
-}
+export 'package:domain/src/counter.dart' show Counter;
