+++ domain/pubspec.yaml
@@ -5,3 +5,5 @@ homepage:
 
 environment:
   sdk: '>=3.0.2 <4.0.0'
+dev_dependencies:
+  test: ^1.24.3
